+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Implement upload image api [#10](https://github.com/rokwire/content-building-block/issues/10)
 
+## [1.0.7] - 2021-09-15
+- Implement upload image api [#10](https://github.com/rokwire/content-building-block/issues/10)
+- Implement Twitter feed api [#11](https://github.com/rokwire/content-building-block/issues/11)
 ## [1.0.6] - 2021-06-23
 ### Fixed
 - GET /student_guides results "null" response on missing data in the table [#8](https://github.com/rokwire/content-building-block/issues/8)